requirements.txt
attrs==19.3.0
paho-mqtt==1.5.0
precise-runner==0.3.1
rhasspy-hermes==0.1.7
